clean up a bit

+++ hello_backup.py
@@ -8,9 +8,16 @@ from datetime import datetime, date
 from werkzeug.security import generate_password_hash, check_password_hash
 from wtforms.widgets import TextArea
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user, current_user
+from webforms import LoginForm, UserForm, PostForm, NamerForm, PasswordForm, SearchForm
+from flask_ckeditor import CKEditor
+from werkzeug.utils import secure_filename
+import uuid as uuid
+import os
 
 #Create an instance of Flask
 app = Flask(__name__)
+# CK Editor
+ckeditor = CKEditor(app)
 # Add database
 
 
@@ -19,10 +26,14 @@ app = Flask(__name__)
 
 # MYSQL DB
 # app.config['SQLALCHEMY_DATABASE_URI'] = "mysql://username:password@localhost/db_name"
-app.config['SQLALCHEMY_DATABASE_URI'] = "mysql+pymysql://root:root@localhost/our_users"
+# app.config['SQLALCHEMY_DATABASE_URI'] = "mysql+pymysql://root:root@localhost/our_users"
+app.config['SQLALCHEMY_DATABASE_URI'] = "postgresql+psycopg2://uqugffbuzdsofw:9c7fe3728694b3987efc5142cd9842b8c22652dc15f46f104883e47e2e7d7dbf@ec2-3-230-24-12.compute-1.amazonaws.com:5432/d3e1v5cm24b6qs"
 
 # Secret Key!!
 app.config['SECRET_KEY'] = "super_secret_key"
+
+UPLOAD_FOLDER = 'static/images'
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
 # Initialize database
 db = SQLAlchemy(app)
@@ -42,23 +53,7 @@ def load_user(user_id):
 
 
 # ---------------------------------------------------------
-# Create a Blog Post model
-class Posts(db.Model):
-	id = db.Column(db.Integer, primary_key=True)
-	title = db.Column(db.String(225))
-	content = db.Column(db.Text)
-	author = db.Column(db.String(55))
-	date_posted = db.Column(db.DateTime, default=datetime.utcnow)
-	slug = db.Column(db.String(55))
 
-
-# Create a Post Form
-class PostForm(FlaskForm):
-	title = StringField("Title", validators=[DataRequired()])
-	content = TextAreaField("Content", validators=[DataRequired()])
-	author =StringField("Author", validators=[DataRequired()])
-	slug = StringField("Slug", validators=[DataRequired()])
-	submit = SubmitField()
 
 #Add Post Page
 @app.route('/add-post', methods=['GET', 'POST'])
@@ -67,13 +62,14 @@ def add_post():
 	form = PostForm()
 
 	if form.validate_on_submit():
+		poster = current_user.id
 		post = Posts(title= form.title.data,
 			content = form.content.data,
-			author = form.author.data,
+			poster_id = poster,
 			slug = form.slug.data)
 		form.title.data = ''
 		form.content.data = ''
-		form.author.data = ''
+		# form.author.data = ''
 		form.slug.data = ''
 
 		# Add data to database
@@ -109,7 +105,7 @@ def edit_post(id):
 	form = PostForm()
 	if form.validate_on_submit():
 		post.title = form.title.data
-		post.author = form.author.data
+		# post.author = form.author.data
 		post.slug = form.slug.data
 		post.content = form.content.data
 		# Update database
@@ -117,34 +113,45 @@ def edit_post(id):
 		db.session.commit()
 		flash("Post Has Been Updated")
 		return redirect(url_for('post', id = post.id))
-	form.title.data = post.title
-	form.author.data = post.author
-	form.slug.data = post.slug
-	form.content.data = post.content
-	return render_template('edit_post.html', form = form)
+	if current_user.id == post.poster_id or id == 1:
+		form.title.data = post.title
+		# form.author.data = post.author
+		form.slug.data = post.slug
+		form.content.data = post.content
+		return render_template('edit_post.html', form = form)
+	else:
+		flash("You Aren't Authorzed to Edit This Post")
+		posts = Posts.query.order_by(Posts.date_posted)
+		return render_template("posts.html", posts = posts)
 
 # Delete Post
 @app.route('/posts/delete/<int:id>')
+@login_required
 def delete_post(id):
 	post_to_delete = Posts.query.get_or_404(id)
+	id = current_user.id
+	if id == post_to_delete.poster.id or id == 1:
+		try:
+			db.session.delete(post_to_delete)
+			db.session.commit()
 
-	try:
-		db.session.delete(post_to_delete)
-		db.session.commit()
+			#Return Message
+			flash("Blog Post was DELETED")
 
-		#Return Message
-		flash("Blog Post was DELETED")
-
-		posts = Posts.query.order_by(Posts.date_posted)
-		return render_template("posts.html", posts = posts)
-	except:
-		# Return Error message
-		flash("There was a problem deleting post. Try again....")
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts = posts)
+		except:
+			# Return Error message
+			flash("There was a problem deleting post. Try again....")
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts = posts)
+	else:
+		flash("You are not authorized to Delete That Post. Try again....")
 		posts = Posts.query.order_by(Posts.date_posted)
 		return render_template("posts.html", posts = posts)
 
 # ------------------------------------------------------------
-
+##########################################
 # JSON Thing
 @app.route('/date')
 def get_current_data():
@@ -155,59 +162,8 @@ def get_current_data():
 	}
 	return favorite_pizza
 	return {"Date": date.today()}
-
+#########################################
 # ---------------------------------------------------------------------------------------
-
-# Create Model
-class Users(db.Model, UserMixin):
-	id = db.Column(db.Integer, primary_key=True)
-	username = db.Column(db.String(20), nullable=False, unique=True)
-	name = db.Column(db.String(200), nullable = False)
-	email = db.Column(db.String(200), nullable= False, unique = True)
-	favorite_color = db.Column(db.String(120))
-	date_added = db.Column(db.DateTime, default=datetime.utcnow)
-	# Do some password stuff
-	password_hash = db.Column(db.String(500))
-
-
-	@property
-	def password(self):
-		raise AttributeError('password is not a readable attribute')
-
-	@password.setter
-	def password(self, password):
-		self.password_hash = generate_password_hash(password)
-
-	def verify_password(self, password):
-		return check_password_hash(self.password_hash, password)
-
-	def __repr__(self):
-		return '<Name %r>' % self.name
-
-
-# Create a Form Class
-class UserForm(FlaskForm):
-	name = StringField("Name", validators=[DataRequired()])
-	username = StringField("Username", validators=[DataRequired()])
-	email = StringField("Email", validators=[DataRequired()])
-	favorite_color = StringField("Favorite Color")
-	password_hash = PasswordField('Password', validators=[DataRequired(), EqualTo('password_hash2', message='Passwords Must Match')])
-	password_hash2 = PasswordField('Confirm Password', validators=[DataRequired()])
-	submit = SubmitField('Submit')
-
-
-# Create a Form Class
-class NamerForm(FlaskForm):
-	name = StringField("What's your name", validators=[DataRequired()])
-	submit = SubmitField('Submit')
-
-# Create a Passowrd Form Class
-class PasswordForm(FlaskForm):
-	email = StringField("What's your Email", validators=[DataRequired()])
-	password_hash = PasswordField("What's your Password", validators=[DataRequired()])
-	submit = SubmitField('Submit')
-
-
 
 
 
@@ -216,16 +172,6 @@ class PasswordForm(FlaskForm):
 
 # Create a route decorator
 @app.route('/')
-# def index():
-# 	return "<h1>Hello World</h1>"
-# Filters
-# safe
-# capitalize
-# lower
-# upper
-# title
-# trim
-# striptags
 def index():
 
 	first_name = "John"
@@ -235,11 +181,15 @@ def index():
 	return render_template("index.html", first_name = first_name,
 		stuff = stuff,
 		favorite_pizza = favorite_pizza)
+
+#################################
 # -----------------------------------------------------------------------
+################################
 # localhost:5000/user/john
 @app.route('/user/<name>')
 def user(name):
 	return render_template("user.html", user_name = name)
+###################################
 # ----------------------------------------------------------------------
 
 # Create Custom Error Page
@@ -255,7 +205,7 @@ def page_not_found(e):
 	return render_template("500.html"), 500
 
 # ----------------------------------------------------------------------------
-
+#########################
 # Create Name Page
 @app.route('/name', methods=['GET', 'POST'])
 def name():
@@ -270,11 +220,12 @@ def name():
 	return render_template('name.html', 
 		name = name, 
 		form = form)
-
+##################################
 # -------------------------------------------------------------------------------------
 
-# Update Database Record
+# Update Database Record of the User
 @app.route('/update/<int:id>', methods=['POST', 'GET'])
+@login_required
 def update(id):
 	form = UserForm()
 	name_to_update = Users.query.get_or_404(id)
@@ -300,8 +251,11 @@ def update(id):
 				form = form,
 				name_to_update = name_to_update,
 				id = id)
+
+
 # ------------------------------------------------------------------------------------------
 
+# Add User
 @app.route('/user/add', methods=['GET', 'POST'])
 def add_user():
 	name = None
@@ -329,21 +283,24 @@ def add_user():
 
 # Delete Users
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-	user_to_delete = Users.query.get_or_404(id)
-	name = None
-	form = UserForm()
-	try:
-		db.session.delete(user_to_delete)
-		db.session.commit()
-		flash("User Deleted Successfully")
-		our_users = Users.query.order_by(Users.date_added)
-		return render_template("add_user.html", form=form, name=name, our_users = our_users)
-	except:
-		flash("Whoops!!! There was a problem deleting the user")
-		return render_template("add_user.html", form=form, name=name, our_users = our_users)
-
-
+	if id == current_user:
+		user_to_delete = Users.query.get_or_404(id)
+		name = None
+		form = UserForm()
+		try:
+			db.session.delete(user_to_delete)
+			db.session.commit()
+			flash("User Deleted Successfully")
+			our_users = Users.query.order_by(Users.date_added)
+			return render_template("add_user.html", form=form, name=name, our_users = our_users)
+		except:
+			flash("Whoops!!! There was a problem deleting the user")
+			return render_template("add_user.html", form=form, name=name, our_users = our_users)
+	else:
+		flash("Sorry, You can't delete that user!")
+		return redirect(url_for('dashboard'))
 
 # --------------------------------------------
 
@@ -380,11 +337,6 @@ def test_pw():
 
 
 # -------------------------------------------------------------------
-#Create Login Form
-class LoginForm(FlaskForm):
-	username = StringField("Username", validators=[DataRequired()])
-	password = PasswordField("Password", validators=[DataRequired()])
-	submit = SubmitField("Submit")
 
 
 # Create a Login Page
@@ -407,35 +359,58 @@ def login():
 	return render_template('login.html', form=form)
 
 
-# Create a Login Page
+# Create Dashboard Page
 @app.route('/dashboard', methods=['GET', 'POST'])
 @login_required
 def dashboard():
 	form = UserForm()
 	id = current_user.id
 	name_to_update = Users.query.get_or_404(id)
-	if request.method == 'POST':
+	if request.method == "POST":
 		name_to_update.name = request.form['name']
-		name_to_update.username = request.form['username']
 		name_to_update.email = request.form['email']
 		name_to_update.favorite_color = request.form['favorite_color']
-		try:
+		name_to_update.username = request.form['username']
+		name_to_update.about_author = request.form['about_author']
+		
+		# Check for profile pic
+		if request.files['profile_pic']:
+			name_to_update.profile_pic = request.files['profile_pic']
+
+			# Grab Image Name
+			pic_filename = secure_filename(name_to_update.profile_pic.filename)
+			# Set UUID
+			pic_name = str(uuid.uuid1()) + "_" + pic_filename
+			# Save That Image
+			saver = request.files['profile_pic']
+			
+
+			# Change it to a string to save to db
+			name_to_update.profile_pic = pic_name
+			try:
+				db.session.commit()
+				saver.save(os.path.join(app.config['UPLOAD_FOLDER'], pic_name))
+				flash("User Updated Successfully!")
+				return render_template("dashboard.html", 
+					form=form,
+					name_to_update = name_to_update)
+			except:
+				flash("Error!  Looks like there was a problem...try again!")
+				return render_template("dashboard.html", 
+					form=form,
+					name_to_update = name_to_update)
+		else:
 			db.session.commit()
-			flash("User Updated Successfully")
-			return render_template("dashboard.html",
-				form = form,
+			flash("User Updated Successfully!")
+			return render_template("dashboard.html", 
+				form=form, 
 				name_to_update = name_to_update)
-		except:
-			flash("Error! Looks like there was an error")
-			return render_template("dashboard.html",
-				form = form,
-				name_to_update = name_to_update,
-				id = id)
 	else:
-		return render_template("dashboard.html",
-				form = form,
+		return render_template("dashboard.html", 
+				form=form,
 				name_to_update = name_to_update,
 				id = id)
+
 	return render_template('dashboard.html')
 
 # Create Logout
@@ -445,3 +420,94 @@ def logout():
 	logout_user()
 	flash("You have been Logged Out! Thanks for stopping by")
 	return redirect(url_for('login'))
+
+
+# --------------------------------------------------------------
+
+#Pass Stuff to NavBar
+
+@app.context_processor
+def base():
+	form = SearchForm()
+	return dict(form=form)
+
+
+#Create Search Function
+
+@app.route('/search', methods=['POST'])
+def search():
+	form = SearchForm()
+	posts = Posts.query
+	if form.validate_on_submit():
+		# Get data from submitted form
+		post.searched = form.searched.data
+
+		#Query the database
+		posts = posts.filter(Posts.content.like('%' + post.searched +'%'))
+		posts = posts.order_by(Posts.title).all()
+
+		return render_template('search.html', 
+			form=form, 
+			searched = post.searched,
+			posts = posts)
+
+# -----------------------------------------------------
+
+# Create Admin Page
+
+# Create a route decorator
+@app.route('/admin')
+@login_required
+def admin():
+	id = current_user.id
+	if id == 1:
+		return render_template("admin.html")
+	else:
+		flash("Sorry must be admin to access the Admin page")
+		return redirect(url_for('dashboard'))
+
+# --------------------------------------------------------------
+# Models
+
+# Create a Blog Post model
+class Posts(db.Model):
+	id = db.Column(db.Integer, primary_key=True)
+	title = db.Column(db.String(225))
+	content = db.Column(db.Text)
+	# author = db.Column(db.String(55))
+	date_posted = db.Column(db.DateTime, default=datetime.utcnow)
+	slug = db.Column(db.String(55))
+	# Foreign Key to Link Users (refer to primary key of the user)
+	poster_id = db.Column(db.Integer, db.ForeignKey('users.id'))
+
+
+# Create Model
+class Users(db.Model, UserMixin):
+	id = db.Column(db.Integer, primary_key=True)
+	username = db.Column(db.String(20), nullable=False, unique=True)
+	name = db.Column(db.String(200), nullable = False)
+	email = db.Column(db.String(200), nullable= False, unique = True)
+	favorite_color = db.Column(db.String(120))
+	about_author = db.Column(db.Text(), nullable=True)
+	date_added = db.Column(db.DateTime, default=datetime.utcnow)
+	profile_pic = db.Column(db.String(255), nullable=True)
+	# Do some password stuff
+	password_hash = db.Column(db.String(500))
+	# User Can Have Many Posts
+	posts = db.relationship('Posts', backref='poster')
+
+
+
+	@property
+	def password(self):
+		raise AttributeError('password is not a readable attribute')
+
+	@password.setter
+	def password(self, password):
+		self.password_hash = generate_password_hash(password)
+
+	def verify_password(self, password):
+		return check_password_hash(self.password_hash, password)
+
+	def __repr__(self):
+		return '<Name %r>' % self.name
